Move directory settings from common to directory
