+++ directory/models/__init__.py
@@ -1,4 +1,5 @@
 from directory.models.entry import *  # noqa: F403,F401
 from directory.models.pages import *  # noqa: F403,F401
+from directory.models.settings import *  # noqa: F403,F401
 from directory.models.snippets import *  # noqa: F403,F401
 from directory.models.taxonomy import *  # noqa: F403,F401
