+++ directory/templates/directory/_instance_table_head.html
@@ -5,7 +5,7 @@
 		<th colspan="3" class="instance-table__title">
 			{% trans "Organization" %}
 		</th>
-		{% if settings.common.DirectorySettings.show_scan_results %}
+		{% if settings.directory.DirectorySettings.show_scan_results %}
 			<th class="instance-table__title instance-table__title--security">
 				{% trans "Security" %}
 				{% if page.faq_link %}
