+++ directory/templates/directory/directory_page.html
@@ -102,7 +102,7 @@
 {% endblock main %}
 
 {% block sidebar %}
-	{% if settings.common.DirectorySettings.allow_directory_management %}
+	{% if settings.directory.DirectorySettings.allow_directory_management %}
 		<section class="submit-instance">
 			<h2 class="submit-instance__title">
 				{{ page.submit_title }}
